requirements.txt
asgiref==3.2.10
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-autocomplete-light==3.7.0.dev0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
idna==2.10
irekua-autocomplete==0.1.2
irekua-database==0.1.2
irekua-dev-settings==0.1
irekua-filters==0.1.2
irekua-models==0.1.3
irekua-permissions==0.1.2
irekua-rest-api==0.1.3
itypes==1.2.0
Jinja2==3.0.0a1
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==2.0.0a1
numpy==1.19.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pycountry==20.7.3
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
selia-registration==0.1
selia-templates==0.1.2
selia-visualizers==0.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
timezonefinder==4.4.1
uritemplate==3.0.1
urllib3==1.25.10

